fix($Numbers): Numbers 클래스 실수합 구현
Numbers 클래스에서 정수합을 실수합으로 변경함

before:
public int getSum()
...

after:
public Double getSum() ...

+++ src/main/java/calculator/Model/Numbers.java
@@ -11,10 +11,11 @@ public class Numbers {
         this.numbers = numbers;
     }
 
-    public int getSum() {
+    public Double getSum() {
         return numbers.stream()
-                .mapToInt(Number::getNumber)
+                .mapToDouble(Number::getNumber)
                 .sum();
+
     }
 
     private void validateNumbers(List<Number> numbers) {
